Extract cldl with public setup, improve API and fix security bugs
The motivation for moving it to its own file was that I couldn't get the lcm
trick tests to run.  Here are the major API changes:

1. The CL-group definition and verification of the public setup are separated
from the public key and verification of proofs. In our setting we assume this
had been done beforehand. Now you create a new `CLGroup` with a setup seed.

2. There is no more `q` inside the CL-group definition. Previous code was
very inconsistent about this. Sometimes it used FE::q() and somtimes it would
use the one inside the public key. Now we only use `FE::q`.

Security bugs:

1. Fiat-Shamir challenge space was only 128. Now it's 2^128 as intended.
2. Fiat-Shamir challenge now includes the statement